requirements.txt
django==3.2.14
mysqlclient==2.0.3
pyyaml==6.0
drf-spectacular==0.25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyMySQL==1.0.2
requests==2.28.1
Pillow==9.1.1 #ImageField
channels